Moved the execution of the actionPDFInvoiceRender hook from InvoicePdfGenerator.php to PDFTemplateInvoiceSmarty.php, in order to better handle the invoice data within the template class itself. Also added relevant comments to explain the responsibilities of the different classes involved in the PDF generation process.
